proj/Database/requirements.txt
psycopg2-binary>=2.9.10
pandas>=2.1.4
numpy>=1.26.2
python-dotenv>=1.0.0
pandas-stubs
types-psycopg2
